more stuff

+++ src/main/java/ch/njol/skript/variables/VariablesMap.java
@@ -45,9 +45,9 @@ final class VariablesMap {
 			int i = 0, j = 0;
 			while (i < s1.length() && j < s2.length()) {
 				final char c1 = s1.charAt(i), c2 = s2.charAt(j);
-				if ('0' <= c1 && c1 <= '9' && '0' <= c2 && c2 <= '9') {
+				if ('0' <= c1 && c1 <= '9' && '0' <= c2 && c2 <= '9') { // TODO negative numbers? what about {blah-%number%}? // '-' < '0'
 					final int i2 = StringUtils.findLastDigit(s1, i), j2 = StringUtils.findLastDigit(s2, j);
-					final int n1 = Utils.parseInt("" + s1.substring(i, i2)), n2 = Utils.parseInt("" + s2.substring(j, j2));
+					final long n1 = Utils.parseLong("" + s1.substring(i, i2)), n2 = Utils.parseLong("" + s2.substring(j, j2));
 					if (n1 > n2)
 						return 1;
 					if (n1 < n2)
